tools/requirements/tests.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/tests.txt requirements/tests.in
#
attrs==19.3.0             # via jsonschema, pytest
dataclasses==0.7          # via -r requirements/tools.txt, pydantic
importlib-metadata==1.7.0  # via jsonschema, pluggy, pytest
jsonschema==3.2.0         # via -r requirements/tests.in
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest, pytest-sugar
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pydantic==1.5.1           # via -r requirements/tools.txt
pyparsing==2.4.7          # via packaging
pyrsistent==0.16.0        # via jsonschema
pytest-sugar==0.9.3       # via -r requirements/tests.in
pytest==5.4.3             # via -r requirements/tests.in, pytest-sugar
pyyaml==5.3.1             # via -r requirements/tools.txt
six==1.15.0               # via jsonschema, packaging, pyrsistent
termcolor==1.1.0          # via pytest-sugar
wcwidth==0.2.5            # via pytest
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
